Fix up test cases

+++ com-mmxlabs-lngdataserver/com.mmxlabs.lngdataserver.lng.importers/src/com/mmxlabs/lngdataserver/lng/io/distances/DistancesFromScenarioCopier.java
@@ -81,7 +81,7 @@ public class DistancesFromScenarioCopier {
 				}
 
 				final Route versionRoute = new Route();
-				versionRoute.setDistance(Double.valueOf(l.getDistance()).floatValue());
+				versionRoute.setDistance(Double.valueOf(l.getDistance()));
 				versionRoute.setProvider(l.getProvider());
 				versionRoute.setErrorCode(l.getErrorCode());
 				routes.getRoutes().put(String.format("%s>%s", from, to), versionRoute);
